Fix for enumeration of triples by object
The code had assumed that triples are in object node order in the NQuads file, but in fact they are in line sort order. This means that when enumerating by object, the entire NQuads file has to be scanned rather than being able to terminate the scan when a new object value is seen.